requirements.txt
# --- UI (desktop) ---
PySide6

# --- Core OCR / imaging ---
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
pytesseract==0.3.13
scikit-image==0.24.0
matplotlib==3.8.4

# --- Japanese NLP / normalization ---
mecab-python3==1.0.9
unidic-lite==1.0.8
pykakasi==2.2.1
jaconv==0.3.4
regex==2024.9.11
langdetect==1.0.9

# --- Translation / seq2seq ---
argostranslate==1.9.1
sentencepiece==0.1.99
ctranslate2==3.20.0
sacremoses==0.0.53
stanza==1.1.1

# --- App / persistence / utils ---
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
python-magic==0.4.27
loguru==0.7.2
tqdm==4.66.5
click==8.1.7
pydantic==2.9.2

# --- Flask (optional; safe to keep for legacy web routes) ---
Flask==3.0.3
Werkzeug==3.0.4
Jinja2>=3.1.2
itsdangerous>=2.1.2
blinker>=1.6.2

# --- Dev / test ---
pytest==8.3.3
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2

# --- EasyOCR (optional; needs torch/torchvision) ---
# If you use EasyOCR in your pipeline, keep these lines.
# For Apple Silicon, install CPU wheels for torch/torchvision after activating the venv:
#   pip install --index-url https://download.pytorch.org/whl/cpu torch torchvision
easyocr==1.7.2
scipy==1.16.2
python-bidi==0.6.6
PyYAML==6.0.3
Shapely==2.1.2
pyclipper==1.3.0.post6
ninja==1.13.0
